requirements.txt
aiogram>=3.0.0
python-dotenv>=0.19.0
colorlog~=6.9.0
SQLAlchemy~=2.0.40
python-decouple~=3.8